💣 (#7) - move local state recoil to apollo client var
