+++ src/@types/floatingEmojiType.ts
@@ -1,0 +1,8 @@
+export interface FloatingEmojiType {
+  x: number;
+  y: number;
+  emoji: string;
+  id: string;
+  color: string;
+  name: string;
+}
