+++ src/@types/index.ts
@@ -2,3 +2,4 @@ export * from "./content";
 export * from "./room";
 export * from "./user";
 export * from "./contentType";
+export * from "./floatingEmojiType";
