--- src/atoms/index.ts
@@ -1,2 +1,0 @@
-export { default as floatingEmojisState } from "./floatingEmojis";
-export * from "./floatingEmojis";
