+++ src/components/EmojiEventListener/index.tsx
@@ -1,39 +1,36 @@
 import React, { memo, useMemo } from "react";
 import { useCallback, useEffect } from "react";
-import { useRecoilState } from "recoil";
-import { floatingEmojisState, FloatingEmojiType } from "@atoms";
 import { EmojiEvent, emojiEventName } from "@constant";
 import { FloatingEmoji } from "@components";
 import * as S from "./styles";
 import { v4 } from "uuid";
 import { useEmoji } from "@queries/emoji";
 import { useRoomContext } from "@hooks";
+import { floatingEmojisVar } from "@stores";
+import { useReactiveVar } from "@apollo/client";
+import { FloatingEmojiType } from "@types";
 
 const EmojiEventListener = () => {
-  const [floatingEmojis, setFloatingEmojis] =
-    useRecoilState(floatingEmojisState);
   const room = useRoomContext();
   const { data } = useEmoji(room.code);
+  const floatingEmojis = useReactiveVar(floatingEmojisVar);
 
-  const emojiListener = useCallback(
-    (e: Event) => {
-      if (!(e instanceof EmojiEvent)) {
-        return;
-      }
+  const emojiListener = useCallback((e: Event) => {
+    if (!(e instanceof EmojiEvent)) {
+      return;
+    }
 
-      const newEmoji: FloatingEmojiType = {
-        emoji: e.emoji,
-        id: v4(),
-        x: e.x,
-        y: e.y,
-        color: e.color,
-        name: e.name,
-      };
+    const newEmoji: FloatingEmojiType = {
+      emoji: e.emoji,
+      id: v4(),
+      x: e.x,
+      y: e.y,
+      color: e.color,
+      name: e.name,
+    };
 
-      setFloatingEmojis((prev) => [...prev, newEmoji]);
-    },
-    [setFloatingEmojis],
-  );
+    floatingEmojisVar([...floatingEmojisVar(), newEmoji]);
+  }, []);
 
   useEffect(() => {
     document.addEventListener(emojiEventName, emojiListener);
