+++ src/stores/floatingEmojis.ts
@@ -1,0 +1,6 @@
+import { makeVar } from "@apollo/client";
+import { FloatingEmojiType } from "@types";
+
+const initalState: FloatingEmojiType[] = [];
+
+export const floatingEmojisVar = makeVar<FloatingEmojiType[]>(initalState);
