+++ src/stores/index.ts
@@ -1,3 +1,4 @@
 export * from "./duration";
 export * from "./playTime";
 export * from "./forcePlayTime";
+export * from "./floatingEmojis";
